+++ launchDaemonManagement.zsh
@@ -30,7 +30,7 @@
 export PATH=/usr/bin:/bin:/usr/sbin:/sbin:/usr/local:/usr/local/bin
 
 # Script Version
-scriptVersion="2.4.0"
+scriptVersion="3.0.0a1"
 
 # Client-side Log
 scriptLog="/var/log/org.churchofjesuschrist.log"